+++ src/main/scala/com/github/unchama/seichiassist/subsystems/home/domain/SubHomePersistence.scala
@@ -1,4 +1,4 @@
-package com.github.unchama.seichiassist.subsystems.subhome.domain
+package com.github.unchama.seichiassist.subsystems.home.domain
 
 import cats.{Functor, Monad}
 import com.github.unchama.seichiassist.subsystems.subhome.domain.OperationResult.RenameResult
